[Jit kernel] ci add jit kernel tests
